+++ src/app/components/diabetes-intro/stats.tsx
